requirements.txt
fitz==0.0.1.dev2
matplotlib==3.5.2
opencv_python==4.5.5.64
PyMuPDF==1.18.14
python_pptx==0.6.21
tqdm==4.64.0
